+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - OPI-106: Title field for cards
+- OPI-108: Test outcomes support cards
 
 ### Changed